fix: fix product and cart page navigation
